when user login, go on game list
