+++ mobileapp/App.js
@@ -8,6 +8,7 @@ import * as Font from "expo-font";
 import { Home } from "./components/home/Home";
 import { Game } from "./components/game/Game";
 import { Login } from "./components/user/Login";
+import { HomeGame } from "./components/game/HomeGame";
 
 function cacheFonts(fonts) {
   return fonts.map((font) => Font.loadAsync(font));
@@ -65,6 +66,7 @@ export default function App() {
           />
           <Stack.Screen name="Game" component={Game} />
           <Stack.Screen name="Login" component={Login} />
+          <Stack.Screen name="HomeGame" component={HomeGame} />
         </Stack.Navigator>
       </NavigationContainer>
     </SafeAreaView>
