+++ mobileapp/components/game/GameListItem.tsx
